# docs/openapi/prompts.yaml
openapi: 3.1.0
info:
  title: Prompts & Rendering API
  version: 1.0.0
paths:
  /v1/teams/{teamID}/prompts:
    post:
      summary: Create a Prompt
      description: Creates a new prompt container.
      operationId: createPrompt
      tags:
        - Prompts
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: teamID
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the team.
      x-test-coverage:
        - "TestCreatePrompt_Success: Verifies prompt creation and response payload."
        - "TestCreatePrompt_MissingName: Verifies that missing name returns 400 and 'name is required'."
        - "TestCreatePrompt_Unauthorized: Verifies that unauthenticated request returns 401."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePromptRequest"
      responses:
        "201":
          description: Prompt created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  prompt:
                    $ref: "#/components/schemas/Prompt"
                required:
                  - prompt
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "./auth.yaml#/components/schemas/APIError"
              examples:
                invalid_body:
                  value: { "error": "invalid request body" }
                name_required:
                  value: { "error": "name is required" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "./auth.yaml#/components/schemas/APIError"
              example:
                error: "unauthorized"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "./auth.yaml#/components/schemas/APIError"
              example:
                error: "internal error"

    get:
      summary: List Prompts
      description: Lists all available prompt containers.
      operationId: listPrompts
      tags:
        - Prompts
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: teamID
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the team.
      x-test-coverage:
        - "TestListPrompts: Verifies listing populated prompt containers."
        - "TestListPrompts_Empty: Verifies that listing when empty returns an empty array."
      responses:
        "200":
          description: List of prompts
          content:
            application/json:
              schema:
                type: object
                properties:
                  prompts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Prompt"
                required:
                  - prompts
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "./auth.yaml#/components/schemas/APIError"
              example:
                error: "unauthorized"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "./auth.yaml#/components/schemas/APIError"
              example:
                error: "internal error"
  /v1/prompts/{id}:
    get:
      summary: Get a Prompt
      description: Returns details of a specific prompt container by its unique UUID.
      operationId: getPrompt
      tags:
        - Prompts
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The unique UUID of the prompt.
          schema:
            type: string
            format: uuid
      x-test-coverage:
        - "TestGetPrompt_Success: Verifies finding a prompt by ID."
        - "TestGetPrompt_NotFound: Verifies 404 response on missing prompt ID."
      responses:
        "200":
          description: Prompt details
          content:
            application/json:
              schema:
                type: object
                properties:
                  prompt:
                    $ref: "#/components/schemas/Prompt"
                required:
                  - prompt
        "400":
          description: Invalid UUID format
          content:
            application/json:
              schema:
                $ref: "./auth.yaml#/components/schemas/APIError"
              example:
                error: "invalid prompt id"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "./auth.yaml#/components/schemas/APIError"
              example:
                error: "unauthorized"
        "404":
          description: Prompt not found
          content:
            application/json:
              schema:
                $ref: "./auth.yaml#/components/schemas/APIError"
              example:
                error: "prompt not found"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "./auth.yaml#/components/schemas/APIError"
              example:
                error: "internal error"

    delete:
      summary: Delete a Prompt
      description: Deletes a specific prompt container along with all its template versions.
      operationId: deletePrompt
      tags:
        - Prompts
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The unique UUID of the prompt to delete.
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Prompt successfully deleted
        "400":
          description: Invalid UUID format
          content:
            application/json:
              schema:
                $ref: "./auth.yaml#/components/schemas/APIError"
              example:
                error: "invalid prompt id"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "./auth.yaml#/components/schemas/APIError"
              example:
                error: "unauthorized"
        "404":
          description: Prompt not found
          content:
            application/json:
              schema:
                $ref: "./auth.yaml#/components/schemas/APIError"
              example:
                error: "prompt not found"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "./auth.yaml#/components/schemas/APIError"
              example:
                error: "internal error"
...
components:
  schemas:
    Prompt:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier.
        name:
          type: string
          description: Name of the prompt container.
        description:
          type: string
          description: Brief summary explaining the prompt purpose.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - created_at
        - updated_at
...
    CreatePromptRequest:
      type: object
      properties:
        name:
          type: string
          example: My Prompt
        description:
          type: string
          example: Useful prompt
        team_ids:
          type: array
          items:
            type: string
            format: uuid
      required:
        - name
...